(#17): OverlayButtonGroup 컴포넌트 제거 및 SideButtonGroup으로 대체, GeoTripInitializer에서 오류 메시지 수정
