+++ src/pages/geotrip/components/SideButtonGroup/index.tsx
@@ -1,0 +1,1 @@
+export { default as SideButtonGroup } from './SideButtonGroup';
