+++ src/pages/geotrip/components/index.ts
@@ -3,4 +3,3 @@ export { default as TourSlide } from './TourSlide';
 export { default as GeoTripInitializer } from './GeoTripInitializer';
 export { default as TourResultSwiper } from './TourResultSwiper';
 export { default as TourDetail } from './TourDetail';
-export { default as OverlayButtonGroup } from './OverlayButtonGroup';
