+++ src/pages/geotrip/service/index.ts
@@ -1,1 +1,2 @@
 export { default as useGeoLocationBasedTourQuery } from './getLocationBasedData';
+export { default as usePollySpeechMutation } from './pollyTTS';
